+++ SintezPPKinCharacteristics/Function.h
@@ -36,7 +36,7 @@ public:
 
 	static int									sign( char c );
 	static Function_p							create( const NS_CORE InternalGearRatioValue& k, const NS_CORE W& w, const NS_CORE N& n, const NS_CORE GearSetNumber gearSet );
-	virtual float								calc( float ) const override;
+	virtual double								calc(double) const override;
 
 };
